+++ lib/compose/src/main/kotlin/org/florisboard/lib/compose/FlorisButtons.kt
@@ -66,7 +66,7 @@ fun FlorisButton(
     colors: ButtonColors = ButtonDefaults.buttonColors(),
 ) {
     Button(
-        modifier = modifier.defaultMinSize(minHeight = 44.dp),
+        modifier = modifier.defaultMinSize(minHeight = FlorisTouchTarget.MinSize),
         enabled = enabled,
         shape = shape,
         colors = colors,
@@ -103,7 +103,7 @@ fun FlorisOutlinedButton(
     colors: ButtonColors = ButtonDefaults.outlinedButtonColors(),
 ) {
     OutlinedButton(
-        modifier = modifier.defaultMinSize(minHeight = 44.dp),
+        modifier = modifier.defaultMinSize(minHeight = FlorisTouchTarget.MinSize),
         enabled = enabled,
         shape = shape,
         colors = colors,
@@ -140,7 +140,7 @@ fun FlorisTextButton(
     colors: ButtonColors = ButtonDefaults.textButtonColors(),
 ) {
     TextButton(
-        modifier = modifier.defaultMinSize(minHeight = 44.dp),
+        modifier = modifier.defaultMinSize(minHeight = FlorisTouchTarget.MinSize),
         enabled = enabled,
         shape = shape,
         colors = colors,
@@ -203,7 +203,7 @@ fun FlorisIconButton(
     colors: IconButtonColors = IconButtonDefaults.iconButtonColors(),
     interactionSource: MutableInteractionSource? = null,
     shape: Shape = MaterialTheme.shapes.small,
-    size: DpSize = DpSize(44.dp, 44.dp),
+    size: DpSize = DpSize(FlorisTouchTarget.MinSize, FlorisTouchTarget.MinSize),
     content: @Composable () -> Unit,
 ) {
     // This is a modified version of https://cs.android.com/androidx/platform/frameworks/support/+/androidx-main:compose/material3/material3/src/commonMain/kotlin/androidx/compose/material3/IconButton.kt;l=235-266;drc=e6d33dd5d0a60001a5784d84123b05308d35f410